+++ src/com/intellij/idea/plugin/hybris/project/configurators/GradleConfigurator.kt
@@ -19,7 +19,6 @@
 package com.intellij.idea.plugin.hybris.project.configurators
 
 import com.intellij.idea.plugin.hybris.project.descriptors.impl.GradleModuleDescriptor
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.Project
 import org.jetbrains.plugins.gradle.service.project.open.linkAndRefreshGradleProject
 
@@ -34,7 +33,4 @@ class GradleConfigurator {
             .forEach { linkAndRefreshGradleProject(it, project) }
     }
 
-    companion object {
-        fun getInstance(): GradleConfigurator? = ApplicationManager.getApplication().getService(GradleConfigurator::class.java)
-    }
 }