+++ src/com/intellij/idea/plugin/hybris/project/configurators/LibRootsConfigurator.kt
@@ -22,7 +22,6 @@ package com.intellij.idea.plugin.hybris.project.configurators
 
 import com.intellij.idea.plugin.hybris.project.descriptors.ModuleDescriptor
 import com.intellij.idea.plugin.hybris.project.descriptors.YModuleDescriptor
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.externalSystem.service.project.IdeModifiableModelsProvider
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.roots.ModifiableRootModel
@@ -38,7 +37,4 @@ interface LibRootsConfigurator {
         indicator: ProgressIndicator
     )
 
-    companion object {
-        fun getInstance(): LibRootsConfigurator = ApplicationManager.getApplication().getService(LibRootsConfigurator::class.java)
-    }
 }