+++ src/com/intellij/idea/plugin/hybris/project/configurators/RunConfigurationConfigurator.kt
@@ -18,9 +18,7 @@
  */
 package com.intellij.idea.plugin.hybris.project.configurators
 
-import com.intellij.idea.plugin.hybris.project.configurators.impl.DefaultRunConfigurationConfigurator
 import com.intellij.idea.plugin.hybris.project.descriptors.HybrisProjectDescriptor
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.project.Project
 
@@ -33,7 +31,4 @@ interface RunConfigurationConfigurator {
         cache: HybrisConfiguratorCache
     )
 
-    companion object {
-        val instance: RunConfigurationConfigurator = ApplicationManager.getApplication().getService(DefaultRunConfigurationConfigurator::class.java)
-    }
 }