+++ src/com/intellij/idea/plugin/hybris/project/configurators/impl/DataSourcesConfigurator.kt
@@ -30,7 +30,6 @@ import com.intellij.database.util.DataSourceUtil
 import com.intellij.database.util.DbImplUtil
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.properties.PropertyService
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.module.ModuleManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.OrderRootType
@@ -101,10 +100,6 @@ class DataSourcesConfigurator {
 
         dataSource.resolveDriver()
         dataSource.ensureDriverConfigured()
-    }
-
-    companion object {
-        fun getInstance(): DataSourcesConfigurator? = ApplicationManager.getApplication().getService(DataSourcesConfigurator::class.java)
     }
 
 }